forge fmt, line coverage & gas snapshot

+++ src/utils/DecimalMath.sol
@@ -51,7 +51,11 @@ library DecimalMath {
      * (x * UNIT) * (y * UNIT) = x * y * UNIT ^ 2,
      * the result is divided by UNIT to remove double scaling.
      */
-    function mulDecimal(uint256 x, uint256 y) internal pure returns (uint256 z) {
+    function mulDecimal(uint256 x, uint256 y)
+        internal
+        pure
+        returns (uint256 z)
+    {
         return (x * y) / UNIT;
     }
 
@@ -62,7 +66,11 @@ library DecimalMath {
      * (x * UNIT) / (y * UNIT) = x / y (Decimal representation is lost),
      * x is first scaled up to end up with a decimal representation.
      */
-    function divDecimal(uint256 x, uint256 y) internal pure returns (uint256 z) {
+    function divDecimal(uint256 x, uint256 y)
+        internal
+        pure
+        returns (uint256 z)
+    {
         return (x * UNIT) / y;
     }
 
@@ -72,7 +80,11 @@ library DecimalMath {
      * E.g. if value is not a decimal, a scale up by 18 makes it a low precision decimal.
      * If value is a low precision decimal, a scale up by 9 makes it a high precision decimal.
      */
-    function upscale(uint x, uint factor) internal pure returns (uint) {
+    function upscale(uint256 x, uint256 factor)
+        internal
+        pure
+        returns (uint256)
+    {
         return x * 10 ** factor;
     }
 
@@ -84,7 +96,11 @@ library DecimalMath {
      *
      * Scaling down a regular integer would not make sense.
      */
-    function downscale(uint x, uint factor) internal pure returns (uint) {
+    function downscale(uint256 x, uint256 factor)
+        internal
+        pure
+        returns (uint256)
+    {
         return x / 10 ** factor;
     }
 
@@ -97,28 +113,44 @@ library DecimalMath {
     /**
      * @dev See mulDecimal for uint256.
      */
-    function mulDecimalUint128(uint128 x, uint128 y) internal pure returns (uint128) {
+    function mulDecimalUint128(uint128 x, uint128 y)
+        internal
+        pure
+        returns (uint128)
+    {
         return (x * y) / UNIT_UINT128;
     }
 
     /**
      * @dev See divDecimal for uint256.
      */
-    function divDecimalUint128(uint128 x, uint128 y) internal pure returns (uint128) {
+    function divDecimalUint128(uint128 x, uint128 y)
+        internal
+        pure
+        returns (uint128)
+    {
         return (x * UNIT_UINT128) / y;
     }
 
     /**
      * @dev See upscale for uint256.
      */
-    function upscaleUint128(uint128 x, uint factor) internal pure returns (uint128) {
+    function upscaleUint128(uint128 x, uint256 factor)
+        internal
+        pure
+        returns (uint128)
+    {
         return x * (10 ** factor).to128();
     }
 
     /**
      * @dev See downscale for uint256.
      */
-    function downscaleUint128(uint128 x, uint factor) internal pure returns (uint128) {
+    function downscaleUint128(uint128 x, uint256 factor)
+        internal
+        pure
+        returns (uint128)
+    {
         return x / (10 ** factor).to128();
     }
 
@@ -143,14 +175,18 @@ library DecimalMath {
     /**
      * @dev See upscale for uint256.
      */
-    function upscale(int x, uint factor) internal pure returns (int) {
+    function upscale(int256 x, uint256 factor) internal pure returns (int256) {
         return x * (10 ** factor).toInt();
     }
 
     /**
      * @dev See downscale for uint256.
      */
-    function downscale(int x, uint factor) internal pure returns (int) {
+    function downscale(int256 x, uint256 factor)
+        internal
+        pure
+        returns (int256)
+    {
         return x / (10 ** factor).toInt();
     }
 
@@ -161,28 +197,44 @@ library DecimalMath {
     /**
      * @dev See mulDecimal for uint256.
      */
-    function mulDecimalInt128(int128 x, int128 y) internal pure returns (int128) {
+    function mulDecimalInt128(int128 x, int128 y)
+        internal
+        pure
+        returns (int128)
+    {
         return (x * y) / UNIT_INT128;
     }
 
     /**
      * @dev See divDecimal for uint256.
      */
-    function divDecimalInt128(int128 x, int128 y) internal pure returns (int128) {
+    function divDecimalInt128(int128 x, int128 y)
+        internal
+        pure
+        returns (int128)
+    {
         return (x * UNIT_INT128) / y;
     }
 
     /**
      * @dev See upscale for uint256.
      */
-    function upscaleInt128(int128 x, uint factor) internal pure returns (int128) {
+    function upscaleInt128(int128 x, uint256 factor)
+        internal
+        pure
+        returns (int128)
+    {
         return x * ((10 ** factor).toInt()).to128();
     }
 
     /**
      * @dev See downscale for uint256.
      */
-    function downscaleInt128(int128 x, uint factor) internal pure returns (int128) {
+    function downscaleInt128(int128 x, uint256 factor)
+        internal
+        pure
+        returns (int128)
+    {
         return x / ((10 ** factor).toInt().to128());
     }
 }
